requirements.txt
discord.py>=2.3
aiosqlite>=0.19
aiohttp>=3.8
